Remove leave page helpers and move functionality into leave page class
