+++ OrangeTests/Pages/LeavePage.cs
@@ -1,14 +1,36 @@
 ﻿using OpenQA.Selenium;
+using OrangeHRMTests.Helpers;
+using OrangeHRMTests.Extensions;
 
 namespace OrangeHRMTests.Locators
 {
     public class LeavePage
     {
         private readonly IWebDriver _driver;
+        private readonly GlobalHelpers _globalHelpers;
+        private readonly GlobalLocators _globalLocators;
+        private readonly Dictionary<int, string> monthsAsStrings = new()
+            {
+                { 1, "January" },
+                { 2, "February" },
+                { 3, "March" },
+                { 4, "April" },
+                { 5, "May" },
+                { 6, "June" },
+                { 7, "July" },
+                { 8, "August" },
+                { 9, "September" },
+                { 10, "October" },
+                { 11, "November" },
+                { 12, "December" },
+            };
 
-        public LeavePage(IWebDriver driver)
+        public LeavePage(IWebDriver driver, GlobalHelpers globalHelpers, GlobalLocators globalLocators)
         {
             _driver = driver;
+            _globalHelpers = globalHelpers;
+            _globalLocators = globalLocators;
+
         }
 
         public string InsufficientBalanceErrorText = "Balance not sufficient";
@@ -42,5 +64,142 @@ namespace OrangeHRMTests.Locators
         public List<IWebElement> DropdownOptions => DropdownListBox.FindElements(By.XPath(".//div[@role='option']")).Skip(1).ToList();
         public List<IWebElement> RemoveStatusButtons => _driver.FindElements(By.XPath("//form//i[contains(@class, 'bi-x')]")).ToList();
         public IWebElement EmployeeNameInputField => _driver.FindElement(By.XPath("//label[contains(., 'Employee Name')]//parent::div//following-sibling::div//input"));
+
+        
+
+        public void SelectRandomLeaveType()
+        {
+            _globalHelpers.Wait.Until(d => LeaveTypeSelectElement.Displayed);
+            LeaveTypeSelectElement.Click();
+            _globalHelpers.SelectRandomElement(DropdownOptions);
+        }
+
+        public (DateTime startDate, DateTime endDate) GetRandomLeaveDates(int duration)
+        {
+            var currentDate = DateTime.Now.Date;
+            var startDate = currentDate.AddDays(new Random().Next(10, 31));
+            startDate = EnsureWorkingDay(startDate);
+            var endDate = startDate.AddDays(duration);
+            endDate = EnsureWorkingDay(endDate);
+
+            return (startDate, endDate);
+        }
+
+        private DateTime EnsureWorkingDay(DateTime initialDate)
+        {
+            var adjustedDate = initialDate.Date;
+            if (adjustedDate.DayOfWeek == DayOfWeek.Saturday) adjustedDate = adjustedDate.AddDays(2);
+            else if (adjustedDate.DayOfWeek == DayOfWeek.Sunday) adjustedDate = adjustedDate.AddDays(1);
+
+            return adjustedDate;
+        }
+
+        public void SelectDate(DateTime date)
+        {
+            _globalHelpers.Wait.Until(d => MonthSelector.Displayed);
+            MonthSelector.ClickViaJavaScript();
+            _globalHelpers.Wait.Until(d => MonthsWrapper.Displayed);
+            Months.SelectItemByText(monthsAsStrings[date.Month]);
+            _globalHelpers.ClickViaActions(YearSelector);
+            _globalHelpers.Wait.Until(d => YearWrapper.Displayed);
+            Years.SelectItemByText(date.Year.ToString());
+            Dates.SelectItemByText(date.Day.ToString());
+        }
+
+        public List<IWebElement> FindRecordsForDateRangeAndStatus(DateTime startDate, DateTime endDate, string leaveStatus, bool matchEmployeeName = false)
+        {
+            var matchingRecords = new List<IWebElement>();
+            foreach (var record in LeaveRecords)
+            {
+                var recordData = _globalHelpers.GetRowCells(record);
+                (var fromDate, var toDate) = ParseRecordDates(recordData[1]);
+                if (!(fromDate == startDate && toDate == endDate)) continue;
+                if (!recordData[6].Text.Contains(leaveStatus)) continue;
+                if (matchEmployeeName)
+                {
+                    var firstAndLastName = _globalLocators.UserName.Text.Split(' ');
+                    if (!recordData[2].Text.Contains(firstAndLastName[0]) &&
+                        !recordData[2].Text.Contains(firstAndLastName[1])) continue;
+                }
+
+                matchingRecords.Add(record);
+            }
+
+            return matchingRecords;
+        }
+
+        private (DateTime fromDate, DateTime? toDate) ParseRecordDates(IWebElement datesElement)
+        {
+            var dates = datesElement.Text.Split(" to ");
+            var fromDateStrings = dates[0].Split('-');
+            var fromDateAsString = $"{fromDateStrings[2]}/{fromDateStrings[1]}/{fromDateStrings[0]}";
+            var fromDate = DateTime.Parse(fromDateAsString);
+            DateTime? toDate = null;
+
+            if (dates.Length > 1)
+            {
+                var toDateStrings = dates[1].Split('-');
+                var toDateAsString = $"{toDateStrings[2]}/{toDateStrings[1]}/{toDateStrings[0]}";
+                toDate = DateTime.Parse(toDateAsString);
+            }
+
+            return (fromDate, toDate);
+        }
+
+        public void ApplyForLeave(DateTime startDate, DateTime endDate)
+        {
+            NavigateToLeavePage();
+            SelectRandomLeaveType();
+
+            // Select Start Date
+            FromDateInputField.SendKeys(startDate.ToString(_globalHelpers.dateFormatString));
+
+            // Select End Date
+            ToDateInputField.ClearViaSendKeys();
+            ToDateInputField.SendKeys(endDate.ToString(_globalHelpers.dateFormatString));
+            ToDateInputField.SendKeys(Keys.Tab);
+
+            _globalHelpers.Wait.Until(d => PartialDaysSelectElement.Displayed);
+            PartialDaysSelectElement.Click();
+            PartialDaysOptions.SelectItemByText("All Days");
+            _globalHelpers.Wait.Until(d => DurationSelectElement.Displayed);
+            DurationSelectElement.Click();
+            DurationOptions.SelectItemByText("Half Day - Morning");
+            CommentsTextArea.SendKeys("Personal leave/vacation");
+            ApplyButton.Submit();
+            _globalHelpers.Wait.Until(d => _globalLocators.SuccessAlert.Displayed);
+        }
+
+        private void NavigateToLeavePage()
+        {
+            _globalHelpers.LoginAs("admin", true);
+            _globalHelpers.Wait.Until(d => _globalLocators.UserDropdown.Displayed);
+            _globalLocators.LeaveLink.Click();
+            _globalHelpers.Wait.Until(d => ApplyLink.Displayed);
+            ApplyLink.Click();
+            _globalHelpers.Wait.Until(d => ApplyButton.Displayed);
+        }
+
+        public void AssertRecordExistsWithStatus(bool recordExists, IWebElement leaveRecord, string leaveStatus, string expectedStatus)
+        {
+            Assert.Multiple(() =>
+            {
+                Assert.That(recordExists, Is.True);
+                Assert.That(leaveStatus, Is.EqualTo(expectedStatus));
+            });
+        }
+
+        public void GotoLeaveList()
+        {
+            LeaveListLink.Click();
+            _globalHelpers.Wait.Until(d => _globalLocators.RecordsTable.Displayed);
+        }
+
+        public void CancelLeave()
+        {
+            CancelLeaveButton.Click();
+            _globalHelpers.Wait.Until(d => _globalLocators.SuccessAlert.Displayed);
+            _globalHelpers.Wait.Until(d => LeaveRecords.Count > 0);
+        }
     }
 }
