+++ OrangeTests/Tests/LeavePageTests.cs
@@ -1,6 +1,5 @@
 ﻿using OpenQA.Selenium;
 using OpenQA.Selenium.Chrome;
-using OrangeHRMTests.Extensions;
 using OrangeHRMTests.Helpers;
 using OrangeHRMTests.Locators;
 
@@ -12,7 +11,6 @@ namespace OrangeHRMTests.Tests
         private GlobalHelpers _globalHelpers;
         private GlobalLocators _globalLocators;
         private LeavePage _leavePage;
-        private LeavePageHelpers _leavePageHelpers;
 
         [SetUp]
         public void Setup()
@@ -20,8 +18,7 @@ namespace OrangeHRMTests.Tests
             _driver = new ChromeDriver();
             _globalLocators = new GlobalLocators(_driver);
             _globalHelpers = new GlobalHelpers(_driver, new Random(), _globalLocators);
-            _leavePage = new LeavePage(_driver);
-            _leavePageHelpers = new LeavePageHelpers(_leavePage, _globalHelpers, _globalLocators);
+            _leavePage = new LeavePage(_driver, _globalHelpers, _globalLocators);
         }
 
         //[TestCase(1)]
@@ -29,28 +26,28 @@ namespace OrangeHRMTests.Tests
         //[TestCase(14)]
         public void CanApplyForLeave(int duration)
         {
-            (var startDate, var endDate) = _leavePageHelpers.GetRandomLeaveDates(duration);
+            (var startDate, var endDate) = _leavePage.GetRandomLeaveDates(duration);
             var mustMatchEmployeeName = true;
-            _leavePageHelpers.ApplyForLeave(startDate, endDate);
+            _leavePage.ApplyForLeave(startDate, endDate);
 
             // Check that record exists in main Leave List and My Leave list.
-            _leavePageHelpers.GotoLeaveList();
-            var leaveRecords = _leavePageHelpers.FindRecordsForDateRangeAndStatus(startDate, endDate, "Pending", mustMatchEmployeeName);
+            _leavePage.GotoLeaveList();
+            var leaveRecords = _leavePage.FindRecordsForDateRangeAndStatus(startDate, endDate, "Pending", mustMatchEmployeeName);
             Assert.That(leaveRecords.Count, Is.EqualTo(1));
 
             _leavePage.MyLeaveLink.Click();
             _globalHelpers.Wait.Until(d => _globalLocators.RecordsTable.Displayed);
-            leaveRecords = _leavePageHelpers.FindRecordsForDateRangeAndStatus(startDate, endDate, "Pending", mustMatchEmployeeName);
+            leaveRecords = _leavePage.FindRecordsForDateRangeAndStatus(startDate, endDate, "Pending", mustMatchEmployeeName);
             Assert.That(leaveRecords.Count, Is.EqualTo(1));
 
             // Cancel leave and confirm it now shows as cancelled
             var numCancelledLeaveRecordsForDateRange =
-                _leavePageHelpers.FindRecordsForDateRangeAndStatus(startDate, endDate, "Cancelled").Count;
-            _leavePageHelpers.CancelLeave();
-            leaveRecords = _leavePageHelpers.FindRecordsForDateRangeAndStatus(startDate, endDate, "Cancelled");
+                _leavePage.FindRecordsForDateRangeAndStatus(startDate, endDate, "Cancelled").Count;
+            _leavePage.CancelLeave();
+            leaveRecords = _leavePage.FindRecordsForDateRangeAndStatus(startDate, endDate, "Cancelled");
             Assert.That(leaveRecords.Count, Is.GreaterThan(numCancelledLeaveRecordsForDateRange));
-            _leavePageHelpers.GotoLeaveList();
-            leaveRecords = _leavePageHelpers.FindRecordsForDateRangeAndStatus(startDate, endDate, "Pending", mustMatchEmployeeName);
+            _leavePage.GotoLeaveList();
+            leaveRecords = _leavePage.FindRecordsForDateRangeAndStatus(startDate, endDate, "Pending", mustMatchEmployeeName);
             Assert.That(leaveRecords.Count, Is.EqualTo(0));
         }
